Create identify an industrial motion system using experimental data
To identify dynamics, the response of the system was measured in a closed loop. Block diagram of the control system was given. A disturbance signal 𝑓 are used and signals 𝑢, 𝑥, 𝑒 were measured. The obtained measurements are included in the file ClosedLoopSignals.mat. The data file contains signals in the order of [f, u, x, e]. A zero reference was used to measure the provided data. The sampling frequency used is given as 8000 Hz