Create Compliant Mechanism Design
When keys or a metal object drop into a gully, they are
most of the time retrievable using a hook or a magnet. This
is not the case with plastic cards. This research focuses on the
gripping part and leaves the rotating part out of the scope. From
three concepts, one is chosen and developed. The design is then
modelled theoretically by doing kinematics and kinetic analysis
on the mechanism. A pseudo-rigid body model and a finite
element analysis are then applied to obtain theoretical results
from the design. From the theoretical model a prototype was
made and tested to obtain data to compare with the theoretical
data. The PRB Model and the visualization is done using MATLAB